% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_null}
\alias{create_null}
\title{Create null objects}
\usage{
create_null(names)
}
\arguments{
\item{names}{Vector of names to be turned into null objects}
}
\value{
Null objects saved into the working environment
}
\description{
This function creates null objects from a vector of names
}
