% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seasonal_mean.R
\name{get_group_mean}
\alias{get_group_mean}
\title{Get group mean}
\usage{
get_group_mean(
  ltm_path,
  fname,
  shapefiles = ecodata::epu_sf \%>\% filter(EPU != "SS"),
  region_name,
  anom = TRUE
)
}
\arguments{
\item{ltm_path}{The file path to the long term mean data}

\item{fname}{The file path to the raster file to use as starting input}

\item{shapefiles}{Shape files for the regions to mask to. Defaults to EPUs.}

\item{region_name}{The names of the regions to calculate}

\item{anom}{Whether to calculate seasonal anomaly}
}
\value{
A data frame
}
\description{
This function gets the group mean by season
}
