% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nc_to_raster}
\alias{nc_to_raster}
\title{Convert netcdf to raster}
\usage{
nc_to_raster(nc, varnum)
}
\arguments{
\item{nc}{The nc object}

\item{varnum}{The index of the variable to convert to a raster}
}
\value{
A raster (cropped to Northeast US)
}
\description{
This function converts a netcdf object to a raster object
}
