% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_sst_anomaly}
\alias{new_sst_anomaly}
\title{SST anomaly using long-term mean from 1991-2020}
\format{
A data frame with 513 rows and 4 variables:
\describe{
\item{\code{Year}}{double Year}
\item{\code{EPU}}{character Ecological Production Unit}
\item{\code{Var}}{character Indicator name}
\item{\code{Value}}{double Anomaly value (degrees C)}
}
}
\usage{
new_sst_anomaly
}
\description{
SST anomaly using long-term mean from 1991-2020
}
\details{
DETAILS
}
\keyword{datasets}
