% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seasonal_mean.R
\name{seasonal_epu_ltm}
\alias{seasonal_epu_ltm}
\title{Get long term seasonal mean}
\usage{
seasonal_epu_ltm(
  ltm,
  shapefiles = ecodata::epu_sf \%>\% filter(EPU != "SS"),
  region_name
)
}
\arguments{
\item{ltm}{The long term temperature data (raster)}

\item{shapefiles}{Shape files for the regions to mask to}

\item{region_name}{The names of the regions to calculate}

\item{anom}{Whether to calculate seasonal anomaly}
}
\value{
A raster (cropped to Northeast US)
}
\description{
This function gets the long term seasonal mean
}
